project v1 done

+++ models/requestBdd.js
@@ -55,41 +55,56 @@ module.exports = {
 
     // INSERT TODO
     insertTodo(userId, message, completion){
-        let sql = "INSERT INTO todos (todoId, message, completion, createdAt, updatedAt, userId) VALUES (NULL, '"+message+"', '"+completion+"', NOW(), 'none', "+userId+")"
-        connexion.query(sql, function (err, result) {
-            if(err){
-                return err
-            }
-            else{
-                console.log("--> INSERT Todo")
-            }
-        })
+        if(userId != "" || message != "" || completion != ""){
+            let sql = "INSERT INTO todos (todoId, message, completion, createdAt, updatedAt, userId) VALUES (NULL, '"+message+"', '"+completion+"', NOW(), 'none', "+userId+")"
+            connexion.query(sql, function (err, result) {
+                if(err){
+                    return err
+                }
+                else{
+                    console.log("--> INSERT Todo")
+                }
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
     },
 
     // UPDATE TODO
     updateTodo(id, completion){
-        let sql = "UPDATE todos SET completion='"+completion+"', updatedAt=NOW() WHERE todoId="+id;
-        connexion.query(sql, function (err, result) {
-            if(err){
-                return err
-            }
-            else{
-                console.log("--> UPDATE Todo")
-            }
-        })
+        if(id != "" || completion != ""){
+            let sql = "UPDATE todos SET completion='"+completion+"', updatedAt=NOW() WHERE todoId="+id;
+            connexion.query(sql, function (err, result) {
+                if(err){
+                    return err
+                }
+                else{
+                    console.log("--> UPDATE Todo")
+                }
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
     },
 
     // DELETE TODO
     deleteTodo(id){
-        let sql = "DELETE FROM todos WHERE todoId="+id
-        connexion.query(sql, function (err, result) {
-            if(err){
-                return err
-            }
-            else{
-                console.log("--> DELETE Todo")
-            }
-        })
+        if(id != ""){
+            let sql = "DELETE FROM todos WHERE todoId="+id
+            connexion.query(sql, function (err, result) {
+                if(err){
+                    return err
+                }
+                else{
+                    console.log("--> DELETE Todo")
+                }
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
     },
     
 
@@ -167,48 +182,80 @@ module.exports = {
         })
     },
 
-    // UPDATE USER
-    updateUser(id, firstname, lastname, username, password, email){
-        bcrypt.hash(password, null, null, function(err, hash) {
-            let password_hash = hash
-            let sql = "UPDATE users SET firstname='"+firstname+"', lastname='"+lastname+"', username='"+username+"', password='"+password_hash+"', email='"+email+"', updatedAt=NOW() WHERE userId="+id
-            connexion.query(sql, function (err, result) {
-                if(err){
-                    return err
-                }
-                else{
-                    console.log("--> UPDATE USER")
-                }
-            })
-        })
-    },
-
-    // INSERT USER
-    insertUser(firstname, lastname, username, password, email){
-        bcrypt.hash(password, null, null, function(err, hash) {
-            let password_hash = hash
-            let sql = "INSERT INTO users (userId, firstname, lastname, username, password, email, createdAt, updatedAt) VALUES (NULL, '"+firstname+"', '"+lastname+"', '"+username+"', '"+password_hash+"', '"+email+"', NOW(), 'none')";
-            connexion.query(sql, function (err, result) {
-                if(err){
-                    return err
-                }
-                else{
-                    console.log("--> INSERT USER")
-                }
-            })
-        })
-    },
-
-    // DELETE USER
-    deleteUser(id){
-        let sql = "DELETE FROM users WHERE userId="+id;
-        connexion.query(sql, function (err, result) {
+    // SELECT FOR SELECTBOX
+    selectBoxUser(res){
+        let sql = "SELECT * FROM users"
+        connexion.query(sql, function (err, result, fields) {
             if(err){
                 return err
             }
             else{
-                console.log("--> DELETE USER")
+                res.render('todos/edit',{
+                    title: 'Add Todo',
+                    todos: "add",
+                    selectBox: result
+                  }) 
             }
         })
+    },
+
+    // UPDATE USER
+    updateUser(id, firstname, lastname, username, password, email){
+        if(id != "" || firstname != "" || lastname != "" || username != "" || password != "" || email != ""){
+            bcrypt.hash(password, null, null, function(err, hash) {
+                let password_hash = hash
+                let sql = "UPDATE users SET firstname='"+firstname+"', lastname='"+lastname+"', username='"+username+"', password='"+password_hash+"', email='"+email+"', updatedAt=NOW() WHERE userId="+id
+                connexion.query(sql, function (err, result) {
+                    if(err){
+                        return err
+                    }
+                    else{
+                        console.log("--> UPDATE USER")
+                    }
+                })
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
+    },
+
+    // INSERT USER
+    insertUser(firstname, lastname, username, password, email){
+        if(firstname != "" || lastname != "" || username != "" || password != "" || email != ""){
+            bcrypt.hash(password, null, null, function(err, hash) {
+                let password_hash = hash
+                let sql = "INSERT INTO users (userId, firstname, lastname, username, password, email, createdAt, updatedAt) VALUES (NULL, '"+firstname+"', '"+lastname+"', '"+username+"', '"+password_hash+"', '"+email+"', NOW(), 'none')";
+                connexion.query(sql, function (err, result) {
+                    if(err){
+                        return err
+                    }
+                    else{
+                        console.log("--> INSERT USER")
+                    }
+                })
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
+    },
+
+    // DELETE USER
+    deleteUser(id){
+        if(id != ""){
+            let sql = "DELETE FROM users WHERE userId="+id;
+            connexion.query(sql, function (err, result) {
+                if(err){
+                    return err
+                }
+                else{
+                    console.log("--> DELETE USER")
+                }
+            })
+        }
+        else{
+            console.log("Veuillez remplir corretement tous les champs")
+        }
     }
 }
